Employee Database
Here is a database I think we can use for our app. It allows us to add and delete employees. It also uses key value pairs so I am hoping it will be easy to implement with what we already have.  When you get a chance to look at it let me know what you think!